backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
ldap3==2.9.1
aiosmtplib==3.0.1
jinja2==3.1.4
reportlab==4.2.2
httpx==0.27.0
email-validator==2.2.0
python-dateutil==2.9.0.post0
